YG-160📂 게시글 수정 파일 분리 및 삭제 후 router 변경
