+++ apis/auth/oauthApi.ts
@@ -3,8 +3,8 @@ import { removeAccessToken, setAccessToken } from "./token/access.utils"
 import { getRefreshToken, removeRefreshToken, setRefreshToken } from "./token/refresh.utils"
 
 const AUTH_API_URL = "/auth"
-// const redirect_uri = "http://localhost:3000/auth"
-const redirect_uri = "https://yeogi-client.vercel.app/auth"
+const redirect_uri = "http://localhost:3000/auth"
+// const redirect_uri = "https://yeogi-client.vercel.app/auth"
 
 /**
  * @function
