+++ app/(afterLogin)/post/edit/[postId]/_components/type.ts
@@ -1,0 +1,6 @@
+export type PostFormProps = {
+    postId: number
+    resetAll: () => void
+    isFreeForm: boolean
+    setIsOverlayOpen: (isOverlayOpen: boolean) => void
+}
